refactor: KEEP-1587 delete old job-spawner and workflow-runner scripts
Remove scripts/scheduler/job-spawner.ts, scripts/runtime/workflow-runner.ts,
and scripts/runtime/workflow-runner-bootstrap.ts/mjs. All superseded by
keeperhub-executor/. Update test paths and spec references to point
to the new locations.
