+++ keeperhub-executor/in-process.ts
@@ -119,7 +119,7 @@ async function updateScheduleStatus(
 
 /**
  * Execute a workflow in-process (no K8s Job).
- * Refactored from scripts/runtime/workflow-runner.ts main() to be callable
+ * Refactored from keeperhub-executor/workflow-runner.ts main() to be callable
  * from the executor without managing its own process lifecycle.
  */
 // biome-ignore lint/complexity/noExcessiveCognitiveComplexity: orchestrates multiple phases of workflow execution
