--- scripts/runtime/workflow-runner-bootstrap.mjs
@@ -1,17 +1,0 @@
-/**
- * ESM Bootstrap script for workflow-runner
- *
- * This ESM script registers a hook to return empty module for 'server-only'
- * before loading the main workflow-runner script.
- *
- * Usage: node --import ./scripts/runtime/workflow-runner-bootstrap.mjs scripts/runtime/workflow-runner.ts
- *
- * Or with tsx:
- *   NODE_OPTIONS="--import ./scripts/runtime/workflow-runner-bootstrap.mjs" npx tsx scripts/runtime/workflow-runner.ts
- */
-
-import { register } from "node:module";
-import { pathToFileURL } from "node:url";
-
-// Register a custom loader to handle 'server-only' imports
-register("./server-only-loader.mjs", pathToFileURL("./"));
