+++ specs/k8s-job-quick-reference.md
@@ -60,8 +60,8 @@ resources:
 
 | File | Purpose |
 |------|---------|
-| `scripts/scheduler/job-spawner.ts` | Polls SQS, creates K8s Jobs |
-| `scripts/runtime/workflow-runner.ts` | Executes workflow in Job pod |
+| `keeperhub-executor/index.ts` | Polls SQS, creates K8s Jobs |
+| `keeperhub-executor/workflow-runner.ts` | Executes workflow in Job pod |
 | `deploy/local/workflow-runner-job.yaml` | Job template |
 
 ## Commands
