+++ specs/schedule-trigger.md
@@ -71,13 +71,13 @@ All services in Minikube - suitable for production-like testing (~8GB RAM).
    - Calls KeeperHub API to execute workflows directly
    - Handles retries and error logging
 
-3. **Job Spawner** (`scripts/scheduler/job-spawner.ts`) - Hybrid/K8s Mode
+3. **Job Spawner** (`keeperhub-executor/index.ts`) - Hybrid/K8s Mode
    - Runs as a Kubernetes Deployment
    - Polls SQS queue for workflow trigger messages
    - Creates K8s Jobs for each workflow execution
    - Provides isolated execution environment
 
-4. **Workflow Runner** (`scripts/runtime/workflow-runner.ts`) - Hybrid/K8s Mode
+4. **Workflow Runner** (`keeperhub-executor/workflow-runner.ts`) - Hybrid/K8s Mode
    - Runs inside K8s Jobs (one per workflow execution)
    - Executes workflow steps in isolation
    - Updates execution status in database
@@ -455,8 +455,8 @@ For production, use the full K8s mode or hybrid mode with:
 | **Scripts** ||
 | `scripts/scheduler/schedule-dispatcher.ts` | Dispatcher script (queries DB, sends to SQS) |
 | `scripts/scheduler/.ts` | Executor script - dev mode (polls SQS, calls API) |
-| `scripts/scheduler/job-spawner.ts` | Job spawner - hybrid/k8s mode (polls SQS, creates K8s Jobs) |
-| `scripts/runtime/workflow-runner.ts` | Workflow runner - runs inside K8s Jobs |
+| `keeperhub-executor/index.ts` | Job spawner - hybrid/k8s mode (polls SQS, creates K8s Jobs) |
+| `keeperhub-executor/workflow-runner.ts` | Workflow runner - runs inside K8s Jobs |
 | **Services** ||
 | `lib/schedule-service.ts` | Schedule management service |
 | `lib/db/schema/workflow-schedules.ts` | Database schema |
