+++ tests/e2e/vitest/full-pipeline.test.ts
@@ -281,7 +281,7 @@ describe.skipIf(SKIP_INFRA_TESTS)("Full Pipeline E2E", () => {
 
       const scriptPath = join(
         import.meta.dirname,
-        "../../../scripts/runtime/workflow-runner-bootstrap.ts"
+        "../../../keeperhub-executor/workflow-runner-bootstrap.ts"
       );
       const child = spawn("npx", ["tsx", scriptPath], {
         env,
